+++ server/app.js
@@ -8,6 +8,7 @@ const expressValidator = require("express-validator");
 // allows us to use .env
 require("dotenv").config();
 // import routes
+const authRoutes = require("./routes/auth");
 const userRoutes = require("./routes/user");
 
 // app
@@ -28,6 +29,7 @@ app.use(cookieParser());
 app.use(expressValidator());
 
 // routes middleware
+app.use("/api", authRoutes);
 app.use("/api", userRoutes);
 
 const port = process.env.PORT || 3000;
