+++ server/routes/user.js
@@ -1,11 +1,17 @@
 const express = require("express");
 const router = express.Router();
 
-const { signup, signin, signout } = require("../controllers/user");
-const { userSignupValidator } = require("../validator");
-// post method for signing up
-router.post("/signup", userSignupValidator, signup);
-router.post("/signin", signin);
-router.get("/signout", signout);
+const { requireSignin, isAuth, isAdmin } = require("../controllers/auth");
+
+const { userById } = require("../controllers/user");
+
+router.get("/secret/:userId", requireSignin, isAuth, isAdmin, (req, res) => {
+  res.json({
+    user: req.profile
+  });
+});
+
+// anytime there is a userid the router will run the method userById
+router.param("userId", userById);
 
 module.exports = router;
